# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ORN-AI Talent Infrastructure Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidates/{id}/cv-file:
    post:
      operationId: uploadCvFile
      tags: [candidates]
      summary: Upload a real CV file (PDF/DOC/DOCX) and auto-extract profile data
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CvFileUploadRequest"
      responses:
        "200":
          description: Updated candidate with extracted profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CvFileUploadResult"
  /candidates/{id}/cv/full.pdf:
    get:
      operationId: downloadFullCv
      tags: [candidates]
      summary: Download the candidate's transformed full CV as a branded PDF
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: PDF
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /candidates/{id}/cv/masked.pdf:
    get:
      operationId: downloadMaskedCv
      tags: [candidates]
      summary: Download an anonymised, recruiter-shareable PDF (no PII)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: PDF
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /candidates/{id}/shortlist:
    post:
      operationId: shortlistCandidate
      tags: [recruiter]
      summary: Toggle the recruiter shortlist flag on a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShortlistRequest"
      responses:
        "200":
          description: Updated candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
...
  /candidates/{id}/skill-gap:
    get:
      operationId: getSkillGap
      tags: [evaluation]
      summary: Skill gap analysis for a candidate against their target role
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Skill gap result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillGap"
  /candidates/{id}/projects:
    get:
      operationId: listCandidateProjects
      tags: [projects]
      summary: List projects assigned to a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Project" }
    post:
      operationId: assignCandidateProject
      tags: [projects]
      summary: Assign a new capstone project to a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project status, feedback or end date
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /candidates/{id}/evaluation:
    post:
      operationId: runEvaluation
      tags: [evaluation]
      summary: Run AI evaluation against a candidate's CV
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Evaluation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evaluation"
    get:
      operationId: getEvaluation
      tags: [evaluation]
      summary: Get the latest evaluation for a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Evaluation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evaluation"
...
components:
  schemas:
...
    EnglishLevel:
      type: string
      enum: [A1, A2, B1, B2, C1, C2]
    VisaStatus:
      type: string
      enum: [eu_citizen, work_permit, blue_card, requires_sponsorship, student_visa]
...
    CvUpload:
      type: object
      properties:
        fileName: { type: string }
        fileSize: { type: integer }
        contentSummary: { type: string }
      required: [fileName, fileSize]
    EvaluationScores:
      type: object
      properties:
        cvQuality: { type: integer, minimum: 0, maximum: 100 }
        technicalSkillMatch: { type: integer, minimum: 0, maximum: 100 }
        technicalRelevance: { type: integer, minimum: 0, maximum: 100 }
        englishReadiness: { type: integer, minimum: 0, maximum: 100 }
        europeJobReadiness: { type: integer, minimum: 0, maximum: 100 }
        marketReadiness: { type: integer, minimum: 0, maximum: 100 }
        careerGapRisk: { type: integer, minimum: 0, maximum: 100 }
        upskillingNeeds: { type: integer, minimum: 0, maximum: 100 }
        overall: { type: integer, minimum: 0, maximum: 100 }
      required:
        - cvQuality
        - technicalSkillMatch
        - technicalRelevance
        - englishReadiness
        - europeJobReadiness
        - marketReadiness
        - careerGapRisk
        - upskillingNeeds
        - overall
    CandidateClassification:
      type: string
      enum: [recruiter_ready, needs_upskilling, needs_reskilling, not_ready_yet]
    SkillGap:
      type: object
      properties:
        targetRole: { type: string }
        required:
          type: array
          items: { type: string }
        matched:
          type: array
          items: { type: string }
        missing:
          type: array
          items: { type: string }
        matchPct: { type: integer, minimum: 0, maximum: 100 }
      required: [targetRole, required, matched, missing, matchPct]
    EvaluationInsight:
      type: object
      properties:
        title: { type: string }
        detail: { type: string }
        severity: { type: string, enum: [strength, opportunity, gap] }
      required: [title, detail, severity]
    Evaluation:
      type: object
      properties:
        candidateId: { type: string }
        evaluatedAt: { type: string, format: date-time }
        scores: { $ref: "#/components/schemas/EvaluationScores" }
        strengths:
          type: array
          items: { type: string }
        gaps:
          type: array
          items: { type: string }
        recommendedUpskilling:
          type: array
          items: { type: string }
        insights:
          type: array
          items: { $ref: "#/components/schemas/EvaluationInsight" }
        readinessTier: { type: string, enum: [emerging, developing, ready, elite] }
        classification: { $ref: "#/components/schemas/CandidateClassification" }
        skillGap: { $ref: "#/components/schemas/SkillGap" }
      required:
        - candidateId
        - evaluatedAt
        - scores
        - strengths
        - gaps
        - recommendedUpskilling
        - insights
        - readinessTier
        - classification
        - skillGap
    Candidate:
      type: object
      properties:
        id: { type: string }
        fullName: { type: string }
        email: { type: string }
        phone: { type: string }
        country: { type: string }
        targetRole: { type: string }
        yearsExperience: { type: integer }
        visaStatus: { $ref: "#/components/schemas/VisaStatus" }
        englishLevel: { $ref: "#/components/schemas/EnglishLevel" }
        euWorkEligible: { type: boolean }
        linkedinUrl: { type: string }
        createdAt: { type: string, format: date-time }
        cv:
          oneOf:
            - $ref: "#/components/schemas/CvUpload"
            - type: "null"
        evaluation:
          oneOf:
            - $ref: "#/components/schemas/Evaluation"
            - type: "null"
        skills:
          type: array
          items: { type: string }
        avatarUrl: { type: string }
        source: { type: string, enum: [direct, recruiter] }
        lastRole:
          oneOf:
            - { type: string }
            - { type: "null" }
        domain:
          oneOf:
            - { type: string }
            - { type: "null" }
        careerGapMonths: { type: integer, minimum: 0 }
        isShortlisted: { type: boolean }
        isClientReady: { type: boolean }
        isIndustryReady: { type: boolean }
        hasCvFile: { type: boolean }
      required:
        - id
        - fullName
        - email
        - phone
        - country
        - targetRole
        - yearsExperience
        - visaStatus
        - englishLevel
        - euWorkEligible
        - linkedinUrl
        - createdAt
        - skills
        - avatarUrl
        - source
        - careerGapMonths
        - isShortlisted
        - isClientReady
        - isIndustryReady
        - hasCvFile
...
    ExtractedProfile:
      type: object
      properties:
        fullName:
          oneOf: [{ type: string }, { type: "null" }]
        email:
          oneOf: [{ type: string }, { type: "null" }]
        phone:
          oneOf: [{ type: string }, { type: "null" }]
        location:
          oneOf: [{ type: string }, { type: "null" }]
        yearsExperience:
          oneOf: [{ type: integer }, { type: "null" }]
        lastRole:
          oneOf: [{ type: string }, { type: "null" }]
        domain:
          oneOf: [{ type: string }, { type: "null" }]
        careerGapMonths: { type: integer, minimum: 0 }
        skills:
          type: array
          items: { type: string }
      required: [careerGapMonths, skills]
    CvFileUploadResult:
      type: object
      properties:
        candidate: { $ref: "#/components/schemas/Candidate" }
        extracted: { $ref: "#/components/schemas/ExtractedProfile" }
      required: [candidate, extracted]
    ProjectStatus:
      type: string
      enum: [in_progress, completed, cancelled]
    Project:
      type: object
      properties:
        id: { type: string }
        candidateId: { type: string }
        name: { type: string }
        techStack:
          type: array
          items: { type: string }
        durationWeeks: { type: integer, minimum: 1, maximum: 52 }
        status: { $ref: "#/components/schemas/ProjectStatus" }
        feedback:
          oneOf: [{ type: string }, { type: "null" }]
        startDate: { type: string, format: date-time }
        endDate:
          oneOf: [{ type: string, format: date-time }, { type: "null" }]
        createdAt: { type: string, format: date-time }
      required: [id, candidateId, name, techStack, durationWeeks, status, startDate, createdAt]
    ProjectCreate:
      type: object
      properties:
        name: { type: string, minLength: 2 }
        techStack:
          type: array
          items: { type: string }
        durationWeeks: { type: integer, minimum: 1, maximum: 52 }
        startDate: { type: string, format: date-time }
      required: [name, techStack, durationWeeks, startDate]
    ProjectUpdate:
      type: object
      properties:
        status: { $ref: "#/components/schemas/ProjectStatus" }
        feedback: { type: string }
        endDate: { type: string, format: date-time }
...
    CvFileUploadRequest:
      type: object
      properties:
        file:
          type: string
          format: binary
      required: [file]
    ShortlistRequest:
      type: object
      properties:
        shortlisted: { type: boolean }
      required: [shortlisted]
...